Resuming work: Working on Rest API Connections and markets
- Added market interface
- Started KuCoin market class
- Created RestAPI interaction to pull kline data

+++ backend/project/Candle.cs
@@ -43,7 +43,7 @@ namespace TradingBot
         public decimal? K { get; set; }
         public decimal? D { get; set; }
 
-        public Candle(int ux, decimal op, decimal hi, decimal lo, decimal cl)
+        public Candle(long ux, decimal op, decimal hi, decimal lo, decimal cl)
         {
             unix = ux;
             open = op;
